+++ CHANGELOG.md
@@ -11,8 +11,10 @@ OTHERS:
 
 - Added support to transform pseudo functions as `transform.scale().translate()`.
 - Improve support to path d pseudo functions as `d.M().L()`.
+- Refactored `d.regularPolygon()` in shape plugin.
 - Redesign and refactored `d.arc()` in shape plugin.
-- Optimized shape plugin lengths with only two decimals.
+- Redesign and refactored `d.barArc()` in shape plugin.
+- Optimized shape plugin lengths with only four decimals.
 
 ## 0.1.0-alpha.3 (26/10/2023)